Reorganized multivariate tests following the package structure
